Refactoring with terrible error messages (C++)
