+++ src/lang/interpreter.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "../rt/behavior.h"
+
 #include <cstddef>
 #include <map>
 #include <memory>
@@ -44,97 +46,25 @@ namespace verona::interpreter
     }
   };
 
-  class Behavior
-  {
-    friend class rt::ui::ScheduleDiagram;
-
-  public:
-    enum class Status
-    {
-      New,
-      Pending,
-      Ready,
-      Running,
-      Done,
-    };
-
-    static std::string status_to_string(Status status)
-    {
-      switch (status)
-      {
-        case Status::New:
-          return "New";
-        case Status::Pending:
-          return "Pending";
-        case Status::Ready:
-          return "Ready";
-        case Status::Running:
-          return "Running";
-        case Status::Done:
-          return "Done";
-        default:
-          return "Unknown";
-      }
-    }
-
-    Status status;
-
-  private:
-    // Static member for naming
-    static int s_behavior_counter;
-
-    // A unique ID, this is used for drawing and naming, it isn't needed for
-    // scheduling.
-    int id;
-
-    // The IDs of the cowns this behavior is waiting on. This is used to create
-    // a better mermaid diagram, it isn't needed for scheduling.
-    std::map<int, rt::objects::DynObject*> ordered_cown;
-
-  public:
-    // The cowns as they were passed in to the cown. These have to be provided
-    // to the new Interpreter to populate the frame
-    std::vector<rt::objects::DynObject*> cowns;
-    // This uses a function object opposed to a Bytecode* to not leak memory
-    rt::objects::DynObject* code;
-    // The number of behaviors that this behavior is waiting on
-    int pred_ctn = 0;
-    // Behaviors which are waiting on this behavior. These will be notified once
-    // this behavior completes
-    std::set<std::shared_ptr<Behavior>> succ;
-
-    Behavior(
-      rt::objects::DynObject* code_,
-      std::vector<rt::objects::DynObject*> cowns_);
-
-    std::string name();
-    std::string id_str();
-
-    Bytecode* spawn();
-    // This completes the behavior by releasing all cowns
-    // decreffing all held objects and informing its successors.
-    void complete();
-  };
-
   class Scheduler
   {
     // All behaviors that are ready to run
-    std::vector<std::shared_ptr<Behavior>> ready = {};
+    std::vector<rt::core::behavior_ptr> ready = {};
     // A map from cowns to the last behavior that is waiting on them.
     //
     // The cowns in the key are weak pointers, they should never be
     // dereferenced.
-    std::unordered_map<rt::objects::DynObject*, std::shared_ptr<Behavior>>
-      cowns = {};
+    std::unordered_map<rt::objects::DynObject*, rt::core::behavior_ptr> cowns =
+      {};
 
   public:
-    void add(std::shared_ptr<Behavior> behavior);
+    void add(rt::core::behavior_ptr behavior);
 
     void start(Bytecode* main);
 
   private:
-    void complete(std::shared_ptr<Behavior> behavior);
+    void complete(rt::core::behavior_ptr behavior);
     void draw_scedule(std::string message);
-    std::shared_ptr<Behavior> get_next();
+    rt::core::behavior_ptr get_next();
   };
 }
